Create directory structure for python tests
Currently the tests are flat in test/py/ as <package>.<test>. With this
patch, the structure is <package>/<test> and <package> is a proper
package (i.e. there is an __init__ file). This allows to run single test
methods selectively for example with nose2. The change is done using
`git mv`, so the `git blame` information is unchanged.